refactor: added utils module for common functionality
